project_labs_1-3/lab2/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9 # For PostgreSQL
neo4j==5.20.0 # For Neo4j
pydantic==2.7.1